requirements.txt
streamlit==1.40.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
mediapipe==0.10.20
pandas==2.2.3